Add a MySQL test database to docker-compose.yml
It can be useful to use a second database for running tests such as Dusk tests. Adding this service to the `docker-compose.yml` file will make it easier to do this.

To use it, uncomment the `mysql_test` service in docker-compose.yml, and then add the following to your `.env.dusk.local` file:
```
DB_HOST=mysql_test
```