Patch issue with environment database password replacement
This update fixes an issue that is caused if you already have a DB_PASSWORD configured in your .env. Prior to this change, if you have this (for example):

```
DB_PASSWORD=mypassword
```

After running `sail:install` it would then look like this:

```
DB_PASSWORD=passwordmypassword
```

And each time you run `sail install` it'll keep growing. This is because only DB_PASSWORD= is being replaced, when it should be replacing the password as well, which can be accomplished using `preg_replace` and a simple regex as in this PR.