Add file extension to start-container for .gitattributes matching
When you publish Sail's Dockerfiles and push them to a git repository, and then checkout that git repository on a Windows machine, the docker application instance will fail to start because the start-container file contains "\r" windows line endings. The common fix for this is to include a .gitattributes file and set .sh files to eol=lf, "*.sh text eol=lf". However, because the start-container script does not include the .sh extension, the script won't be included in gitattributes directives that use "*.sh". This commit fixes that issue by adding the .sh extension to the start-container files.