Run composer install in dev container
Today, when you start a dev container using `curl -s "https://laravel.build/example-app?with=mysql,redis&devcontainer" | bash` (or any version with a dev container I assume), it works great. However, if I push that repo to GitHub and then clone it back down, the dev container will not start.

That happens because the docker-compose.yml refers to ./vendor, which of course doesn't exist (yet).

I have done some testing on my end, and running this command before creating the dev container ensures that `./vendor/laravel/sail/runtimes/8.3(/Dockerfile)` exists before trying to create the container.

The command is from https://laravel.com/docs/11.x/sail#installing-composer-dependencies-for-existing-projects

For this to work, you would need to add the following to the .env file:
```
WWWGROUP=1000
WWWUSER=1000
```

Note: I am completely new to Laravel and somewhat new to Docker. If this is not the fix here, please let me know.